relo'd input
